+++ tests/options.test.ts
@@ -6,11 +6,9 @@ describe('normalizeOptions', () => {
     expect(normalizeOptions()).toEqual({
       linkEntries: [],
       autoLinkEntries: false,
-      autoLinkMinBytes: 3000,
       snippetName: 'vite-style',
       themeRoot: './',
       sourceCodeDir: 'src',
-      templateBudget: undefined,
     })
   })
 
@@ -20,19 +18,15 @@ describe('normalizeOptions', () => {
         snippetName: 'inline-css',
         linkEntries: ['a.css'],
         autoLinkEntries: true,
-        autoLinkMinBytes: 5000,
         themeRoot: 'theme',
         sourceCodeDir: 'app',
-        templateBudget: 40_000,
       }),
     ).toEqual({
       linkEntries: ['a.css'],
       autoLinkEntries: true,
-      autoLinkMinBytes: 5000,
       snippetName: 'inline-css',
       themeRoot: 'theme',
       sourceCodeDir: 'app',
-      templateBudget: 40_000,
     })
   })
 })